api/requirements.txt
flask>=2.3.0
flask-cors>=4.0.0
torch>=2.0.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
pyyaml>=6.0

